+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+# 1.2.0 Credential Manager
+
+* Credential Provider only returns token for TFS/VSTS Links (not everyone)
+
 # 1.1.0 Credential Manager
 
 * Add support for credential managers.